+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  - Add ability to test suite from a url
  - Add ability to test suite from stdin
+ - Add `file` assertion to `stdout` and `stderr`
 
 # v2.3.0